requirements.txt
streamlit
PyPDF2
pinecone
google-generativeai
tqdm
python-dotenv
